refactor(crm): flip @/lib/hubspot -> @/lib/espocrm across 29 files (PR 4 of 5)
Mechanical import flip: every callsite that pulled CRM functions from
src/lib/hubspot.ts now pulls from src/lib/espocrm.ts. Since espocrm.ts
mirrors the 21-function surface exactly (PR #1030), this is a zero-
behavior-change refactor at the TypeScript level.

Files changed (sed in-place):
  src/app/api/admin/analytics/unified/route.ts
  src/app/api/admin/crm/{companies,contacts,deals,owners,pipelines,tickets}/route.ts
  src/app/api/admin/email/{contacts,stats}/route.ts
  src/app/api/admin/leads/route.ts
  src/app/api/admin/pipeline/{board,stats,deals/[id]/move,deals/[id]/notes}/route.ts
  src/app/api/admin/reports/generate/route.ts
  src/app/api/calendar/book/route.ts
  src/app/api/contact/route.ts
  src/app/api/cron/owner-digest/route.ts
  src/app/api/crm/contact/route.ts
  src/app/api/hubspot/ticket/route.ts
  src/app/api/newsletter/send/route.ts
  src/app/api/newsletter-slack/{commands,events}/route.ts
  src/app/api/subscribe/route.ts
  src/app/api/unsubscribe/route.ts
  src/app/api/webhooks/stripe/route.ts
  src/lib/{roi,slack-commands-extra,voice-brief-sources}.ts

NOT touched (intentional):
- src/lib/hubspot.ts itself — left in tree so any code path I missed
  keeps working. Will be removed in PR 5 once 24h of production data
  confirms no other callers.
- src/components/HubSpotScript.tsx + ad-analytics/digest.ts — these
  use NEXT_PUBLIC_HUBSPOT_PORTAL_ID for the marketing analytics tag,
  not the CRM. The HubSpot tracking script stays.

What this gets us live (once deployed):
- /api/contact, /api/subscribe, /api/unsubscribe write to EspoCRM
- Admin /admin/{leads,crm,pipeline,email} pages read from EspoCRM
- /api/cron/owner-digest reports against EspoCRM
- /api/webhooks/stripe upserts Stripe deals into EspoCRM
- The Slack receivers (newsletter-slack/*, espocrm webhook) all line
  up against the same backend

typecheck/lint/format all green. No production change until next deploy.
