+++ src/app/api/admin/analytics/unified/route.ts
@@ -3,7 +3,7 @@ import { requireAdmin } from "@/lib/api-auth";
 import { getConfig } from "@/lib/ssm-config";
 import { getSeoSnapshot } from "@/lib/gsc";
 import { readThrough } from "@/lib/gsc-cache";
-import { isHubSpotConfigured, getPipelineStats, listNewsletterSubscribers } from "@/lib/hubspot";
+import { isHubSpotConfigured, getPipelineStats, listNewsletterSubscribers } from "@/lib/espocrm";
 import { getStripe } from "@/lib/stripe";
 
 async function safeCall<T>(fn: () => Promise<T>): Promise<T | null> {
