+++ src/app/api/admin/crm/pipelines/route.ts
@@ -1,7 +1,7 @@
 import { NextRequest } from "next/server";
 import { requireAdmin } from "@/lib/api-auth";
 import { NextResponse } from "next/server";
-import { getPipelines, isHubSpotConfigured } from "@/lib/hubspot";
+import { getPipelines, isHubSpotConfigured } from "@/lib/espocrm";
 import { mapIntegrationError } from "@/lib/api-errors";
 
 export async function GET(request: NextRequest) {
